requirements.txt
beautifulsoup4==4.12.3
greenlet==3.0.3
numpy==2.1.1
pandas==2.2.3
playwright==1.47.0
pyee==12.0.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
six==1.16.0
soupsieve==2.6
typing_extensions==4.12.2
tzdata==2024.2
